Refactor router utils into CSF library
