+++ addons/docs/package.json
@@ -48,8 +48,8 @@
     "@storybook/addons": "5.3.0-beta.19",
     "@storybook/api": "5.3.0-beta.19",
     "@storybook/components": "5.3.0-beta.19",
+    "@storybook/csf": "5.3.0-beta.19",
     "@storybook/postinstall": "5.3.0-beta.19",
-    "@storybook/router": "5.3.0-beta.19",
     "@storybook/source-loader": "5.3.0-beta.19",
     "@storybook/theming": "5.3.0-beta.19",
     "acorn": "^7.1.0",
