+++ addons/links/package.json
@@ -32,6 +32,7 @@
     "@storybook/addons": "5.3.0-beta.19",
     "@storybook/client-logger": "5.3.0-beta.19",
     "@storybook/core-events": "5.3.0-beta.19",
+    "@storybook/csf": "5.3.0-beta.19",
     "@storybook/router": "5.3.0-beta.19",
     "core-js": "^3.0.1",
     "global": "^4.3.2",
