+++ addons/links/src/preview.ts
@@ -7,7 +7,7 @@ import {
 import qs from 'qs';
 import addons from '@storybook/addons';
 import { STORY_CHANGED, SELECT_STORY } from '@storybook/core-events';
-import { toId } from '@storybook/router/utils';
+import { toId } from '@storybook/csf';
 import { logger } from '@storybook/client-logger';
 
 interface ParamsId {
