+++ addons/storyshots/storyshots-puppeteer/package.json
@@ -30,8 +30,8 @@
   },
   "dependencies": {
     "@hypnosphi/jest-puppeteer-axe": "^1.4.0",
+    "@storybook/csf": "5.3.0-beta.19",
     "@storybook/node-logger": "5.3.0-beta.19",
-    "@storybook/router": "5.3.0-beta.19",
     "@types/jest-image-snapshot": "^2.8.0",
     "core-js": "^3.0.1",
     "jest-image-snapshot": "^2.8.2",
