+++ lib/api/package.json
@@ -29,6 +29,7 @@
     "@storybook/channels": "5.3.0-beta.19",
     "@storybook/client-logger": "5.3.0-beta.19",
     "@storybook/core-events": "5.3.0-beta.19",
+    "@storybook/csf": "5.3.0-beta.19",
     "@storybook/router": "5.3.0-beta.19",
     "@storybook/theming": "5.3.0-beta.19",
     "core-js": "^3.0.1",
