+++ lib/api/src/modules/stories.ts
@@ -1,5 +1,5 @@
 // FIXME: we shouldn't import from dist but there are no types otherwise
-import { toId, sanitize, parseKind } from '@storybook/router';
+import { toId, sanitize, parseKind } from '@storybook/csf';
 import deprecate from 'util-deprecate';
 
 import { Module } from '../index';
