+++ lib/client-api/package.json
@@ -32,7 +32,7 @@
     "@storybook/channels": "5.3.0-beta.19",
     "@storybook/client-logger": "5.3.0-beta.19",
     "@storybook/core-events": "5.3.0-beta.19",
-    "@storybook/router": "5.3.0-beta.19",
+    "@storybook/csf": "5.3.0-beta.19",
     "core-js": "^3.0.1",
     "eventemitter3": "^4.0.0",
     "global": "^4.3.2",
