+++ lib/codemod/package.json
@@ -29,8 +29,8 @@
   "dependencies": {
     "@hypnosphi/jscodeshift": "^0.6.4",
     "@mdx-js/mdx": "^1.5.1",
+    "@storybook/csf": "5.3.0-beta.19",
     "@storybook/node-logger": "5.3.0-beta.19",
-    "@storybook/router": "5.3.0-beta.19",
     "core-js": "^3.0.1",
     "cross-spawn": "^7.0.0",
     "globby": "^10.0.1",
