+++ lib/codemod/src/transforms/csf-to-mdx.js
@@ -1,18 +1,5 @@
 import recast from 'recast';
-
-// FIXME: duplicate code from @storybook/core start.js
-function isExportStory(key, { includeStories, excludeStories }) {
-  function matches(storyKey, arrayOrRegex) {
-    if (Array.isArray(arrayOrRegex)) {
-      return arrayOrRegex.includes(storyKey);
-    }
-    return storyKey.match(arrayOrRegex);
-  }
-  return (
-    (!includeStories || matches(key, includeStories)) &&
-    (!excludeStories || !matches(key, excludeStories))
-  );
-}
+import { isExportStory } from '@storybook/csf';
 
 function exportMdx(root, options) {
   // eslint-disable-next-line no-underscore-dangle
