+++ lib/codemod/src/transforms/storiesof-to-csf.js
@@ -1,6 +1,6 @@
 import prettier from 'prettier';
 import { logger } from '@storybook/node-logger';
-import { storyNameFromExport } from '@storybook/router';
+import { storyNameFromExport } from '@storybook/csf';
 import { sanitizeName } from '../lib/utils';
 
 /**
