+++ lib/core/package.json
@@ -37,6 +37,7 @@
     "@storybook/client-api": "5.3.0-beta.19",
     "@storybook/client-logger": "5.3.0-beta.19",
     "@storybook/core-events": "5.3.0-beta.19",
+    "@storybook/csf": "5.3.0-beta.19",
     "@storybook/node-logger": "5.3.0-beta.19",
     "@storybook/router": "5.3.0-beta.19",
     "@storybook/theming": "5.3.0-beta.19",
