+++ lib/core/src/client/preview/index.js
@@ -1,5 +1,5 @@
 import { ClientApi, StoryStore, ConfigApi } from '@storybook/client-api';
-import { toId } from '@storybook/router/utils';
+import { toId } from '@storybook/csf';
 import start from './start';
 
 export default {
