+++ lib/core/src/client/preview/start.js
@@ -7,7 +7,7 @@ import AnsiToHtml from 'ansi-to-html';
 import addons from '@storybook/addons';
 import createChannel from '@storybook/channel-postmessage';
 import { ClientApi, StoryStore, ConfigApi } from '@storybook/client-api';
-import { toId, storyNameFromExport } from '@storybook/router/utils';
+import { toId, storyNameFromExport, isExportStory } from '@storybook/csf';
 import { logger } from '@storybook/client-logger';
 import Events from '@storybook/core-events';
 
@@ -23,22 +23,6 @@ const classes = {
   NOPREVIEW: 'sb-show-nopreview',
   ERROR: 'sb-show-errordisplay',
 };
-
-function matches(storyKey, arrayOrRegex) {
-  if (Array.isArray(arrayOrRegex)) {
-    return arrayOrRegex.includes(storyKey);
-  }
-  return storyKey.match(arrayOrRegex);
-}
-
-export function isExportStory(key, { includeStories, excludeStories }) {
-  return (
-    // https://babeljs.io/docs/en/babel-plugin-transform-modules-commonjs
-    key !== '__esModule' &&
-    (!includeStories || matches(key, includeStories)) &&
-    (!excludeStories || !matches(key, excludeStories))
-  );
-}
 
 function showMain() {
   document.body.classList.remove(classes.NOPREVIEW);
