+++ lib/core/src/client/preview/start.test.js
@@ -2,7 +2,7 @@
 import { history, document, window } from 'global';
 
 import Events from '@storybook/core-events';
-import start, { isExportStory } from './start';
+import start from './start';
 
 jest.mock('@storybook/client-logger');
 jest.mock('global', () => ({
@@ -141,43 +141,5 @@ describe('STORY_INIT', () => {
     store.emit(Events.STORY_INIT);
     expect(history.replaceState).toHaveBeenCalledWith({}, '', 'pathname?bar=baz&id=kind--story');
     expect(store.setSelection).toHaveBeenCalledWith({ storyId: 'kind--story' });
-  });
-});
-
-describe('story filters for module exports', () => {
-  it('should exclude __esModule', () => {
-    expect(isExportStory('__esModule', {})).toBeFalsy();
-  });
-
-  it('should include all stories when there are no filters', () => {
-    expect(isExportStory('a', {})).toBeTruthy();
-  });
-
-  it('should filter stories by arrays', () => {
-    expect(isExportStory('a', { includeStories: ['a'] })).toBeTruthy();
-    expect(isExportStory('a', { includeStories: [] })).toBeFalsy();
-    expect(isExportStory('a', { includeStories: ['b'] })).toBeFalsy();
-
-    expect(isExportStory('a', { excludeStories: ['a'] })).toBeFalsy();
-    expect(isExportStory('a', { excludeStories: [] })).toBeTruthy();
-    expect(isExportStory('a', { excludeStories: ['b'] })).toBeTruthy();
-
-    expect(isExportStory('a', { includeStories: ['a'], excludeStories: ['a'] })).toBeFalsy();
-    expect(isExportStory('a', { includeStories: [], excludeStories: [] })).toBeFalsy();
-    expect(isExportStory('a', { includeStories: ['a'], excludeStories: ['b'] })).toBeTruthy();
-  });
-
-  it('should filter stories by regex', () => {
-    expect(isExportStory('a', { includeStories: /a/ })).toBeTruthy();
-    expect(isExportStory('a', { includeStories: /.*/ })).toBeTruthy();
-    expect(isExportStory('a', { includeStories: /b/ })).toBeFalsy();
-
-    expect(isExportStory('a', { excludeStories: /a/ })).toBeFalsy();
-    expect(isExportStory('a', { excludeStories: /.*/ })).toBeFalsy();
-    expect(isExportStory('a', { excludeStories: /b/ })).toBeTruthy();
-
-    expect(isExportStory('a', { includeStories: /a/, excludeStories: ['a'] })).toBeFalsy();
-    expect(isExportStory('a', { includeStories: /.*/, excludeStories: /.*/ })).toBeFalsy();
-    expect(isExportStory('a', { includeStories: /a/, excludeStories: /b/ })).toBeTruthy();
   });
 });
