+++ lib/core/src/client/preview/url.js
@@ -1,6 +1,6 @@
 import { history, document } from 'global';
 import qs from 'qs';
-import { toId } from '@storybook/router/utils';
+import { toId } from '@storybook/csf';
 
 export function pathToId(path) {
   const match = (path || '').match(/^\/story\/(.+)/);
