+++ lib/csf/README.md
@@ -1,0 +1,3 @@
+# Storybook CSF
+
+A minimal set of utility functions for dealing with Storybook [Component Story Format](https://storybook.js.org/docs/formats/component-story-format/).
