+++ lib/source-loader/package.json
@@ -30,7 +30,7 @@
   "dependencies": {
     "@storybook/addons": "5.3.0-beta.19",
     "@storybook/client-logger": "5.3.0-beta.19",
-    "@storybook/router": "5.3.0-beta.19",
+    "@storybook/csf": "5.3.0-beta.19",
     "core-js": "^3.0.1",
     "estraverse": "^4.2.0",
     "global": "^4.3.2",
