+++ lib/source-loader/src/server/abstract-syntax-tree/parse-helpers.js
@@ -1,4 +1,4 @@
-const { toId } = require('@storybook/router/utils');
+const { toId } = require('@storybook/csf');
 
 const STORIES_OF = 'storiesOf';
