+++ lib/source-loader/src/server/abstract-syntax-tree/traverse-helpers.js
@@ -1,4 +1,4 @@
-import { storyNameFromExport } from '@storybook/router/utils';
+import { storyNameFromExport } from '@storybook/csf';
 import { handleADD, handleSTORYOF, patchNode, handleExportedName } from './parse-helpers';
 
 const estraverse = require('estraverse');
@@ -23,6 +23,8 @@ export function splitSTORYOF(ast, source) {
 export function splitExports(ast, source) {
   const parts = [];
   let lastIndex = 0;
+  const excludeStories = ['text'];
+  const includeStories = undefined;
 
   estraverse.traverse(ast, {
     fallback: 'iteration',
