+++ lib/ui/paths.js
@@ -11,6 +11,7 @@ module.exports = {
   '@storybook/channels': dirname(resolve('@storybook/channels/package.json')),
   '@storybook/components': dirname(resolve('@storybook/components/package.json')),
   '@storybook/core-events': dirname(resolve('@storybook/core-events/package.json')),
+  '@storybook/csf': dirname(resolve('@storybook/csf/package.json')),
   '@storybook/router': dirname(resolve('@storybook/router/package.json')),
   '@storybook/theming': dirname(resolve('@storybook/theming/package.json')),
   '@storybook/ui': dirname(resolve('@storybook/ui/package.json')),
